ethica-backend/requirements.txt
flask
flask-cors
PyJWT
Flask-APScheduler
werkzeug
python-dotenv
